my work-in-progress personal portfolio webpage
I've been working on this for a couple weeks, trying to play with different color schemes and HTML elements. I initially wanted to put this when it's all finish, but I'd like GitHub to record how often I'm working on this. Hopefully throughout time, I'll be more proficient in committing to a project, so I don't feel the need to only add the code when it's reaches a "finished" point. For now, I learn best by attempting to clone.